---------------------------------------------------------------------- Nicolas Regnault - discover major bug in FactorialCoefficient (problem with power 2 mult and div). Has to be fixed. - change FactorialCoefficient to support improvements due to 64bits architecture.
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@54 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/MathTools/FactorialCoefficient.h
@@ -46,9 +46,9 @@ class FactorialCoefficient
   int DenominatorPosition;
 
   // array representing the numerator
-  int* Numerator;
+  long* Numerator;
   // array representing the denominator
-  int* Denominator;
+  long* Denominator;
 
  public:
 
@@ -59,13 +59,13 @@ class FactorialCoefficient
   // constructor from an integer
   //
   // x = value to assign to the factorial coefficient
-  FactorialCoefficient(int x);  
+  FactorialCoefficient(long x);  
 
   // constructor from a rational number
   //
   // x = numerator to assign to the factorial coefficient
   // y = denominator to assign to the factorial coefficient
-  FactorialCoefficient(int x, int y);  
+  FactorialCoefficient(long x, long y);  
 
   // destructor
   //
@@ -80,51 +80,51 @@ class FactorialCoefficient
   //
   // x = integer to use
   // return value = reference on current coefficient
-  FactorialCoefficient& operator *= (int x);
+  FactorialCoefficient& operator *= (long x);
 
   // divide by an integer
   //
   // y = integer to use
   // return value = reference on current coefficient
-  FactorialCoefficient& operator /= (int y);
+  FactorialCoefficient& operator /= (long y);
 
   // multiply the coefficient by a power of 2
   // 
   // power = power exponent (must be greater than 0)
   // return value = reference on current coefficient
-  FactorialCoefficient& Power2Multiply (int power);
+  FactorialCoefficient& Power2Multiply (long power);
 
   // divide the coefficient by a power of 2
   // 
   // power = power exponent (must be greater than 0)
   // return value = reference on current coefficient
-  FactorialCoefficient& Power2Divide (int power);
+  FactorialCoefficient& Power2Divide (long power);
 
   // multiply the coefficient by the factorial of an integer
   // 
   // x = integer to use
   // return value = reference on current coefficient
-  FactorialCoefficient& FactorialMultiply (int x);
+  FactorialCoefficient& FactorialMultiply (long x);
 
   // multiply the coefficient by the partial factorial end! / (start - 1)!
   // 
   // start = first integer in the partial factorial product
   // end = last integer in the partial factorial product
   // return value = reference on current coefficient
-  FactorialCoefficient& PartialFactorialMultiply (int start, int end);
+  FactorialCoefficient& PartialFactorialMultiply (long start, long end);
 
   // multiply the coefficient by the factorial of an integer
   // 
   // x = integer to use
   // return value = reference on current coefficient
-  FactorialCoefficient& FactorialDivide (int x);
+  FactorialCoefficient& FactorialDivide (long x);
   
   // divide the coefficient by the partial factorial end! / (start - 1)!
   // 
   // start = first integer in the partial factorial product
   // end = last integer in the partial factorial product
   // return value = reference on current coefficient
-  FactorialCoefficient& PartialFactorialDivide (int start, int end);
+  FactorialCoefficient& PartialFactorialDivide (long start, long end);
 
   // return numerical value associated to the coefficient
   //
@@ -134,7 +134,7 @@ class FactorialCoefficient
   // return integer value associated to the coefficient (0 if the coefficient is not an integer, or can't be cast into an integer)
   //
   // return value = numerical value associated to the coefficient
-  int GetIntegerValue();
+  long GetIntegerValue();
 
  private:
 
@@ -143,14 +143,14 @@ class FactorialCoefficient
   // m = first integer  
   // n = second integer (must be greater than m)
   // return value = GCD
-  int FindGCD(int m, int n);
+  long FindGCD(long m, long n);
 
   // find greatest common divider (recurisive part of the method)
   //
   // m = first integer  
   // n = second integer (must be greater than m)
   // return value = GCD
-  int RecursiveFindGCD(int m, int n);
+  long RecursiveFindGCD(long m, long n);
 
 };
